move basic circuits to rounds, make it build
