+++ prover/crates/bin/witness_generator/src/rounds/basic_circuits/utils.rs
@@ -1,183 +1,38 @@
-use std::{
-    collections::HashSet,
-    hash::{DefaultHasher, Hash, Hasher},
-    sync::Arc,
-    time::Instant,
+use crate::precalculated_merkle_paths_provider::PrecalculatedMerklePathsProvider;
+use crate::rounds::basic_circuits::Witness;
+use crate::storage_oracle::StorageOracle;
+use crate::utils::{
+    expand_bootloader_contents, save_circuit, save_ram_premutation_queue_witness,
+    ClosedFormInputWrapper, KZG_TRUSTED_SETUP_FILE,
 };
-
-use async_trait::async_trait;
-use circuit_definitions::{
-    circuit_definitions::base_layer::{ZkSyncBaseLayerCircuit, ZkSyncBaseLayerStorage},
-    encodings::recursion_request::RecursionQueueSimulator,
-    zkevm_circuits::{
-        fsm_input_output::ClosedFormInputCompactFormWitness,
-        scheduler::{
-            block_header::BlockAuxilaryOutputWitness, input::SchedulerCircuitInstanceWitness,
-        },
-    },
+use crate::witness::WitnessStorage;
+use circuit_definitions::circuit_definitions::base_layer::{
+    ZkSyncBaseLayerCircuit, ZkSyncBaseLayerStorage,
 };
+use circuit_definitions::encodings::recursion_request::RecursionQueueSimulator;
+use circuit_definitions::zkevm_circuits::fsm_input_output::ClosedFormInputCompactFormWitness;
+use std::collections::HashSet;
+use std::hash::{DefaultHasher, Hash, Hasher};
+use std::sync::Arc;
 use tokio::sync::Semaphore;
 use tracing::Instrument;
 use zkevm_test_harness::witness::oracle::WitnessGenerationArtifact;
-use zksync_config::configs::FriWitnessGeneratorConfig;
-use zksync_multivm::{
-    circuit_sequencer_api_latest::{
-        boojum::{
-            field::goldilocks::{GoldilocksExt2, GoldilocksField},
-            gadgets::recursion::recursive_tree_hasher::CircuitGoldilocksPoseidon2Sponge,
-        },
-        geometry_config::get_geometry_config,
-    },
-    interface::storage::StorageView,
-    vm_latest::{constants::MAX_CYCLES_FOR_TX, HistoryDisabled, StorageOracle as VmStorageOracle},
-    zk_evm_latest::ethereum_types::Address,
-};
+use zksync_multivm::circuit_sequencer_api_latest::boojum::field::goldilocks::GoldilocksField;
+use zksync_multivm::circuit_sequencer_api_latest::geometry_config::get_geometry_config;
+use zksync_multivm::interface::storage::StorageView;
+use zksync_multivm::vm_latest::constants::MAX_CYCLES_FOR_TX;
+use zksync_multivm::vm_latest::HistoryDisabled;
+use zksync_multivm::vm_latest::StorageOracle as VmStorageOracle;
+use zksync_multivm::zk_evm_latest::ethereum_types::Address;
 use zksync_object_store::ObjectStore;
-use zksync_prover_dal::{ConnectionPool, Prover};
-use zksync_prover_fri_types::{keys::ClosedFormInputKey, CircuitAuxData};
+use zksync_prover_fri_types::keys::ClosedFormInputKey;
+use zksync_prover_fri_types::CircuitAuxData;
 use zksync_prover_interface::inputs::WitnessInputData;
-use zksync_prover_keystore::keystore::Keystore;
 use zksync_system_constants::BOOTLOADER_ADDRESS;
-use zksync_types::{
-    basic_fri_types::AggregationRound, protocol_version::ProtocolSemanticVersion, L1BatchNumber,
-};
-
-use crate::{
-    artifacts::ArtifactsManager,
-    metrics::WITNESS_GENERATOR_METRICS,
-    precalculated_merkle_paths_provider::PrecalculatedMerklePathsProvider,
-    storage_oracle::StorageOracle,
-    utils::{
-        expand_bootloader_contents, save_circuit, save_ram_premutation_queue_witness,
-        ClosedFormInputWrapper, KZG_TRUSTED_SETUP_FILE,
-    },
-    witness::WitnessStorage,
-    witness_generator::JobManager,
-};
-
-mod artifacts;
-pub mod job_processor;
-
-#[derive(Clone)]
-pub struct BasicCircuitArtifacts {
-    pub(super) circuit_urls: Vec<(u8, String)>,
-    pub(super) queue_urls: Vec<(u8, String, usize)>,
-    pub(super) scheduler_witness: SchedulerCircuitInstanceWitness<
-        GoldilocksField,
-        CircuitGoldilocksPoseidon2Sponge,
-        GoldilocksExt2,
-    >,
-    pub(super) aux_output_witness: BlockAuxilaryOutputWitness<GoldilocksField>,
-}
-
-#[derive(Clone)]
-pub struct BasicWitnessGeneratorJob {
-    pub(super) block_number: L1BatchNumber,
-    pub(super) data: WitnessInputData,
-}
-
-#[derive(Debug)]
-pub struct BasicWitnessGenerator {
-    config: Arc<FriWitnessGeneratorConfig>,
-    object_store: Arc<dyn ObjectStore>,
-    public_blob_store: Option<Arc<dyn ObjectStore>>,
-    prover_connection_pool: ConnectionPool<Prover>,
-    protocol_version: ProtocolSemanticVersion,
-}
-
-type Witness = (
-    Vec<(u8, String)>,
-    Vec<(u8, String, usize)>,
-    SchedulerCircuitInstanceWitness<
-        GoldilocksField,
-        CircuitGoldilocksPoseidon2Sponge,
-        GoldilocksExt2,
-    >,
-    BlockAuxilaryOutputWitness<GoldilocksField>,
-);
-
-impl BasicWitnessGenerator {
-    pub fn new(
-        config: FriWitnessGeneratorConfig,
-        object_store: Arc<dyn ObjectStore>,
-        public_blob_store: Option<Arc<dyn ObjectStore>>,
-        prover_connection_pool: ConnectionPool<Prover>,
-        protocol_version: ProtocolSemanticVersion,
-    ) -> Self {
-        Self {
-            config: Arc::new(config),
-            object_store,
-            public_blob_store,
-            prover_connection_pool,
-            protocol_version,
-        }
-    }
-}
-
-#[async_trait]
-impl JobManager for BasicWitnessGenerator {
-    type Job = BasicWitnessGeneratorJob;
-    type Metadata = L1BatchNumber;
-    type Artifacts = BasicCircuitArtifacts;
-
-    async fn process_job(
-        job: BasicWitnessGeneratorJob,
-        object_store: Arc<dyn ObjectStore>,
-        max_circuits_in_flight: Option<usize>,
-        started_at: Instant,
-    ) -> anyhow::Result<BasicCircuitArtifacts> {
-        let BasicWitnessGeneratorJob {
-            block_number,
-            data: job,
-        } = job;
-
-        tracing::info!(
-            "Starting witness generation of type {:?} for block {}",
-            AggregationRound::BasicCircuits,
-            block_number.0
-        );
-
-        let (circuit_urls, queue_urls, scheduler_witness, aux_output_witness) = generate_witness(
-            block_number,
-            object_store,
-            job,
-            max_circuits_in_flight.unwrap(),
-        )
-        .await;
-        WITNESS_GENERATOR_METRICS.witness_generation_time[&AggregationRound::BasicCircuits.into()]
-            .observe(started_at.elapsed());
-        tracing::info!(
-            "Witness generation for block {} is complete in {:?}",
-            block_number.0,
-            started_at.elapsed()
-        );
-
-        Ok(BasicCircuitArtifacts {
-            circuit_urls,
-            queue_urls,
-            scheduler_witness,
-            aux_output_witness,
-        })
-    }
-
-    async fn prepare_job(
-        metadata: L1BatchNumber,
-        object_store: &dyn ObjectStore,
-        _keystore: Keystore,
-    ) -> anyhow::Result<Self::Job> {
-        tracing::info!("Processing FRI basic witness-gen for block {}", metadata.0);
-        let started_at = Instant::now();
-        let job = Self::get_artifacts(&metadata, object_store).await?;
-
-        WITNESS_GENERATOR_METRICS.blob_fetch_time[&AggregationRound::BasicCircuits.into()]
-            .observe(started_at.elapsed());
-
-        Ok(job)
-    }
-}
+use zksync_types::L1BatchNumber;
 
 #[tracing::instrument(skip_all, fields(l1_batch = %block_number))]
-async fn generate_witness(
+pub(super) async fn generate_witness(
     block_number: L1BatchNumber,
     object_store: Arc<dyn ObjectStore>,
     input: WitnessInputData,
