+++ prover/crates/bin/witness_generator/src/rounds/node_aggregation/mod.rs
@@ -7,7 +7,6 @@ use tokio::sync::Semaphore;
 use zkevm_test_harness::witness::recursive_aggregation::{
     compute_node_vk_commitment, create_node_witness,
 };
-use zksync_config::configs::FriWitnessGeneratorConfig;
 use zksync_object_store::ObjectStore;
 use zksync_prover_dal::{ConnectionPool, Prover, ProverDal};
 use zksync_prover_fri_types::{
@@ -19,7 +18,7 @@ use zksync_prover_fri_types::{
         encodings::recursion_request::RecursionQueueSimulator,
         zkevm_circuits::recursion::leaf_layer::input::RecursionLeafParametersWitness,
     },
-    FriProofWrapper,
+    get_current_pod_name, FriProofWrapper,
 };
 use zksync_prover_keystore::{keystore::Keystore, utils::get_leaf_vk_params};
 use zksync_types::{
@@ -30,11 +29,10 @@ use zksync_types::{
 use crate::{
     artifacts::ArtifactsManager,
     metrics::WITNESS_GENERATOR_METRICS,
-    rounds::{JobManager, WitnessGenerator},
+    rounds::JobManager,
     utils::{load_proofs_for_job_ids, save_recursive_layer_prover_input_artifacts},
 };
 mod artifacts;
-mod job_processor;
 
 #[derive(Clone)]
 pub struct NodeAggregationArtifacts {
@@ -63,9 +61,9 @@ pub struct NodeAggregation;
 impl JobManager for NodeAggregation {
     type Job = NodeAggregationWitnessGeneratorJob;
     type Metadata = NodeAggregationJobMetadata;
-    type Artifacts = NodeAggregationArtifacts;
 
     const ROUND: &'static str = "node_aggregation";
+    const SERVICE_NAME: &'static str = "fri_node_aggregation_witness_generator";
 
     #[tracing::instrument(
         skip_all,
@@ -250,7 +248,7 @@ impl JobManager for NodeAggregation {
             .context("failed to acquire DB connection for NodeAggregationWitnessGenerator")?;
         prover_storage
             .fri_witness_generator_dal()
-            .get_node_aggregation_job_attempts(*job_id)
+            .get_node_aggregation_job_attempts(job_id)
             .await
             .map(|attempts| attempts.unwrap_or(0))
             .context("failed to get job attempts for NodeAggregationWitnessGenerator")
@@ -263,11 +261,28 @@ impl JobManager for NodeAggregation {
     ) -> anyhow::Result<()> {
         connection_pool
             .connection()
-            .await
-            .unwrap()
+            .await?
             .fri_witness_generator_dal()
             .mark_node_aggregation_job_failed(&error, job_id)
             .await;
         Ok(())
+    }
+
+    async fn get_metadata(
+        connection_pool: ConnectionPool<Prover>,
+        protocol_version: ProtocolSemanticVersion,
+    ) -> anyhow::Result<Option<(u32, Self::Metadata)>> {
+        let pod_name = get_current_pod_name();
+        let Some(metadata) = connection_pool
+            .connection()
+            .await?
+            .fri_witness_generator_dal()
+            .get_next_node_aggregation_job(protocol_version, &pod_name)
+            .await
+        else {
+            return Ok(None);
+        };
+
+        Ok(Some((metadata.id, metadata)))
     }
 }
